refactor/ 일정 -> 작성자명 필드 삭제

+++ src/main/java/org/example/scheduleappdevelop/schedule/service/ScheduleService.java
@@ -27,20 +27,15 @@ public class ScheduleService {
         if (request.getContent() == null) {
             throw new IllegalArgumentException("내용은 필수값입니다.");
         }
-        if (request.getAuthor() == null) {
-            throw new IllegalArgumentException("작성자명은 필수값입니다.");
-        }
 
         Schedule schedule = new Schedule(
                 request.getTitle(),
-                request.getContent(),
-                request.getAuthor()
+                request.getContent()
         );
 
         Schedule savedSchedule = scheduleRepository.save(schedule);
         return new ScheduleSaveResponse(
                 savedSchedule.getId(),
-                savedSchedule.getAuthor(),
                 savedSchedule.getTitle(),
                 savedSchedule.getContent(),
                 savedSchedule.getCreatedAt(),
@@ -58,7 +53,6 @@ public class ScheduleService {
                     schedule.getId(),
                     schedule.getTitle(),
                     schedule.getContent(),
-                    schedule.getAuthor(),
                     schedule.getCreatedAt(),
                     schedule.getModifiedAt()
             );
@@ -77,7 +71,6 @@ public class ScheduleService {
                 schedule.getId(),
                 schedule.getTitle(),
                 schedule.getContent(),
-                schedule.getAuthor(),
                 schedule.getCreatedAt(),
                 schedule.getModifiedAt()
         );
@@ -100,17 +93,17 @@ public class ScheduleService {
                 schedule.getId(),
                 schedule.getTitle(),
                 schedule.getContent(),
-                schedule.getAuthor(),
                 schedule.getCreatedAt(),
                 schedule.getModifiedAt()
         );
     }
 
     @Transactional
-    public void deleteSchedule(long scheduleId, String password) {
-        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(
-                () -> new IllegalArgumentException("존재하지 않는 일정입니다.")
-        );
+    public void deleteSchedule(long scheduleId) {
+        boolean b = scheduleRepository.existsById(scheduleId);
+        if(!b){
+            throw new IllegalArgumentException("존재하지 않는 일정입니다");
+        }
         scheduleRepository.deleteById(scheduleId);
     }
 }
